useDoubleClick hook modified
